+++ packages/record-webapck-raw-loader/lib/index.ts
@@ -1,0 +1,5 @@
+import recordWebapckLoader from 'record-webapck-loader';
+
+// @ts-ignore
+recordWebapckLoader.raw = true;
+export default recordWebapckLoader;