+++ packages/record-webapck-raw-loader/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 -   init
 
+### Patch Changes
+
+-   Updated dependencies
+    -   record-webapck-loader@2.0.0
+
 ## 1.1.0
 
 ### Minor Changes